PGPRO-5646: remove compiler warning (gcc 11.2.0)
In function ‘ptrack_write_chunk’,
    inlined from ‘ptrackCheckpoint’ at engine.c:397:2:
engine.c:78:13: warning: ‘write’ reading 8 bytes from a region of size 4
[-Wstringop-overread]
   78 |         if (write(fd, chunk, size) != size)
      |             ^~~~~~~~~~~~~~~~~~~~~~
In file included from engine.c:47:
engine.c: In function ‘ptrackCheckpoint’:
engine.h:55:25: note: source object ‘magic’ of size 4
   55 |         char            magic[PTRACK_MAGIC_SIZE];
      |                         ^~~~~
In file included from engine.c:22:
/usr/include/unistd.h:378:16: note: in a call to function ‘write’ declared with
attribute ‘access (read_only, 2, 3)’
  378 | extern ssize_t write (int __fd, const void *__buf, size_t __n) __wur
      |                ^~~~~

In C a pointer to the first field of a structure and a pointer to the structure
itself are always equal. Add a compile-time assertion check in case the field
magic is not the first field in the structure PtrackMapHdr in the future.

Thanks to Maksim Orlov for the review.